api/requirements.txt
fastapi[all]==0.88.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.5
python-jose[cryptography]==3.3.0
sqlmodel==0.0.8
minio==7.2.5
aio-pika==9.4.1
aiofiles==23.2.1
bcrypt==4.1.2
